base64ct: refactor encoder functions
To make it easier to support the "crypt" encoding (which is slightly
different from the others as the alphabet is shifted), this change
refactors encoders to support an arbitrary `encode_6bits` function
per-encoding.
